Revert "Merge branch 'main' into wangela/npm"
This reverts commit 90fecb336ef70803ca640f509581672e0a7dc2e6, reversing
changes made to 744afd4a60f4c319348d1dcbf8e6d123da8cf62c.

+++ android/src/main/java/com/google/android/react/navsdk/CollectionUtil.java
@@ -22,7 +22,7 @@ public class CollectionUtil {
 
   public static int getInt(String name, Map map, int defaultValue) {
     if (map.containsKey(name) && map.get(name) != null) {
-      return ((Double) map.get(name)).intValue();
+      return ((Integer) map.get(name)).intValue();
     }
     return defaultValue;
   }
